requirements.txt
albucore==0.0.13
albumentations==1.4.10
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cx_Freeze==7.2.7
cx_Logging==3.2.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fire==0.7.0
fonttools==4.55.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
joblib==1.4.2
keyboard==0.13.5
kiwisolver==1.4.7
lazy_loader==0.4
lief==0.15.1
lmdb==1.5.1
lxml==5.3.0
matplotlib==3.9.2
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
protobuf==3.20.2
psutil==5.9.8
pyclipper==1.3.0.post6
pydantic==2.10.1
pydantic_core==2.27.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.1.0
tqdm==4.67.0
ttkbootstrap==1.10.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
